+++ packages/xml-views-definition/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ui5-language-assistant/xml-views-definition
 
+## 0.0.3
+
+### Patch Changes
+
+- 5b190da: Fix: Use file uri for definition
+
 ## 0.0.2
 
 ### Patch Changes